rename #29: currentlocation -> lastlocation
